+++ library/lib/hooks/useDragDrop.ts
@@ -1,4 +1,4 @@
-import { dropElementConfig } from "@/constants"
+import { dropElementConfig } from "@/constants/dropElementConfig"
 import { DropNodeData } from "@/types"
 import { generateUUID } from "@/utils"
 import { useReactFlow, type Node } from "@xyflow/react"
